Moved function to newly created DistanceController, for organizing logic
