+++ app/Http/Controllers/BreedStatusController.php
@@ -16,29 +16,6 @@ class BreedStatusController extends Controller
         return $breedArray;
     }
 
-    public function getMilesBetweenZipCodes($zipCodes, $maxMiles, $focusZip)
-    {
-        $client = new \GuzzleHttp\Client();
-        $zipDistanceArray = [];
-        $zipString = '';
-        $index = 0;
-        foreach($zipCodes as $zipCode){
-            if($index == 0){
-                $zipString .= $zipCode;
-            }else{
-                $zipString .= ', ' . $zipCode;
-            }
-            $index++;
-        }
-        $query = "https://www.zipcodeapi.com/rest/" .  env('ZIP_API_KEY')  . "/multi-distance.json" .
-                                "/"         .  $focusZip               . "/"                     .
-                           $zipString       .  "/mile"
-        ;
-        $queryResponse = $client->request('GET', $query);
-        $zipDistanceArray = json_decode($queryResponse ->getBody()->getContents(), true);
-        return $zipDistanceArray['distances'];
-    }
-
     public function getExternalDataForBreed(Request $request)
     {
             $client = new \GuzzleHttp\Client();
